OmniConvert/requirements.txt
streamlit
openai-whisper
googletrans==4.0.0-rc1
fpdf2
python-docx
openpyxl